Bonus problem complete
Create a summary table for the states based on campaign goals. The table consists of goal ranges, the number of times each state occurs and the percentage of projects that were successful, failed, or cancelled per goal range.
Create a line chart which graphs the relationship between a goal's amount and its chances at success, failure, or cancellation.