Prepare charts for use in the analysis report
Add chart and axis titles to all the charts to make it easier to understand what they are visualizing.
Create a Country Distribution pivot chart to visualize the country misrepresentation limitation discussed in Analysis_Report.doc.
Add percentage labels to the Country Distribution bar chart to further emphasize the disproportionality.