Project complete
Create an in-depth analysis report of the conclusions made about Kickstarter campaigns based on the data.
Create a pivot table that displays the percentage of outcomes for each parent category to use in the analysis report.
Create a pivot table to count the number of projects launched each year and visualize the results with a bar chart to use in the analysis report.
Create a new README that gives useful information about the completed project.
Rename the original README.md as Instructions.md and place it and the Images folder into the Instructions folder to keep the instructions available to view in the repository.